+++ src/analyze/src/components/designs/Design.vue
@@ -37,11 +37,18 @@
                 </a>
                 <template v-for="timeframe in join.related_table.timeframes">
                   <a class="panel-block timeframe"
-                      :key="timeframe.label"
                       v-if="!timeframe.hidden"
                       @click="timeframeSelected(timeframe)"
-                      :class="{'is-active': timeframe.selected}">
+                      :key="timeframe.label"
+                      :class="{
+                        'is-active': timeframe.selected,
+                        'is-sqlite-unsupported': isConnectionDialectSqlite(connectionDialect)
+                      }">
                     {{timeframe.label}}
+                    <div class='sqlite-unsupported-container'
+                         v-if='isConnectionDialectSqlite(connectionDialect)'>
+                      <small>Unsupported by SQLite</small>
+                    </div>
                   </a>
                   <template v-if="timeframe.selected">
                     <template v-for="period in timeframe.periods">
@@ -345,6 +352,7 @@ export default {
       'chartsOpen',
       'hasSQLError',
       'sqlErrorMessage',
+      'connectionDialect',
     ]),
     ...mapGetters('designs', [
       'currentModelLabel',
@@ -362,7 +370,12 @@ export default {
       'showJoinColumnAggregateHeader',
       'formattedSql',
     ]),
-
+    ...mapGetters('settings', [
+      'isConnectionDialectSqlite',
+    ]),
+    canToggleTimeframe() {
+      return !this.isConnectionDialectSqlite(this.connectionDialect);
+    },
     limit: {
       get() {
         return this.$store.getters['designs/currentLimit'];
@@ -397,8 +410,10 @@ export default {
       this.$store.dispatch('designs/toggleColumn', column);
       this.$store.dispatch('designs/getSQL', { run: false });
     },
-
     timeframeSelected(timeframe) {
+      if (!this.canToggleTimeframe) {
+        return;
+      }
       this.$store.dispatch('designs/toggleTimeframe', timeframe);
     },
 
@@ -479,6 +494,27 @@ code {
     font-weight: bold;
     &:hover {
       background: white;
+    }
+  }
+
+  &.is-sqlite-unsupported {
+    opacity: .5;
+    cursor: not-allowed;
+    .sqlite-unsupported-container {
+      display: flex;
+      flex-direction: row;
+      justify-content: flex-end;
+      flex-grow: 1;
+
+      small {
+        font-size: 60%;
+        font-style: italic;
+      }
+    }
+    &.timeframe {
+      &::after {
+        display: none;
+      }
     }
   }