+++ tests/fixtures/core.py
@@ -1,5 +1,5 @@
 import pytest
-from unittest import mock
+import os
 
 from meltano.core.project_init_service import ProjectInitService
 from meltano.core.project_add_service import ProjectAddService
@@ -52,6 +52,9 @@ def project(test_dir):
     service = ProjectInitService(PROJECT_NAME)
     project = service.init()
 
+    # this is a test repo, let's remove the `.env`
+    os.unlink(project.root.joinpath(".env"))
+
     # cd into the new project root
-    project.chdir()
+    project.activate()
     return project